+++ xdsl/utils/type.py
@@ -53,26 +53,3 @@ def have_compatible_shape(lhs_type: Attribute, rhs_type: Attribute) -> bool:
         dim1 == DYNAMIC_INDEX or dim2 == DYNAMIC_INDEX or dim1 == dim2
         for dim1, dim2 in zip(lhs_type.get_shape(), rhs_type.get_shape())
     )
-
-
-def are_tosa_broadcastable(t_in1: Attribute, t_in2: Attribute, t_out: Attribute):
-    if (
-        not isinstance(t_in1, ShapedType)
-        or not isinstance(t_in2, ShapedType)
-        or not isinstance(t_out, ShapedType)
-    ):
-        return False
-
-    # check ranks are equal
-    if not (t_in1.get_num_dims() == t_in2.get_num_dims() == t_out.get_num_dims()):
-        return False
-
-    # check ranks are broadcastable
-    in_shapes = zip(t_in1.get_shape(), t_in2.get_shape())
-
-    if not all(dim1 == dim2 or dim1 == 1 or dim2 == 1 for dim1, dim2 in in_shapes):
-        return False
-
-    # check output shape is constructed from input shapes
-    shapes = zip(t_in1.get_shape(), t_in2.get_shape(), t_out.get_shape())
-    return all(dim_out == max(dim1, dim2) for dim1, dim2, dim_out in shapes)
